feat: refactor docgen props.
